decomp: retail-width types for pending externs and struct fields
m2c types globals it cannot infer as s32. Retail knows better: every
absolute data access in the raw-word oracles was decoded (lui/addiu/ori
chains, gp = 0x166C00) and D_XXXXXXXX names carry their address, so each
global's real width/sign/float-ness is known. Two mechanical passes, each
change kept only where the unit's game-compiler score rises strictly:

- extern-width-fix: `extern s32 D_X;` -> u8/s8/u16/s16/f32 in 14 pending
  bodies (e.g. fun_00214970 41.83 -> 73.15 with f32 camera globals,
  fun_0022de10 3.66 -> 23.86, fun_001f79a8 80.00 -> 85.09);
- struct-width-fix: field types of struct M2c_D_XXXXXXXX in 24 per-unit
  headers used only by pending units, padding recomputed so offsets never
  move (set_tfrag_dists 32.43 -> 60.65, fun_001f2c60 33.46 -> 60.14,
  fun_001edaa8 71.18 -> 82.92, update_all_point_lights 49.23 -> 59.24).

draw_mobys_setup's body reads D_0010FA90 as u16 (79.00 -> 80.20). All
changes are in NON_MATCHING bodies/headers; INCLUDE_ASM stays
authoritative. ./verify-baseline.sh: PASS (SHA-256 e0505810...c0c9).
Progress map: C_FUZZY 47.81%.

+++ include/rnc/assembly_textbin_fun_001ed940_types.h
@@ -5,7 +5,7 @@
 
 struct M2c_D_0013F350 {
     u8 pad_0[0x2F0];
-    s32 unk2F0;
+    f32 unk2F0;
     u8 pad_2F4[0xFF0];
     u8 unk12E4;
     u8 unk12E5;
@@ -21,7 +21,7 @@ struct M2c_D_0013F350 {
 
 struct M2c_D_00187080 {
     u8 pad_0[0x8];
-    s32 unk8;
+    f32 unk8;
 };
 
 struct M2c_D_001870D0 {
